requirements.txt
asgiref==3.7.2
black==23.10.0
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
distlib==0.3.7
Django==4.2.6
djangorestframework==3.14.0
drf-yasg==1.21.7
filelock==3.12.4
flake8==6.1.0
identify==2.5.30
idna==3.4
inflection==0.5.1
install==1.3.5
isort==5.12.0
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pre-commit==3.5.0
pycodestyle==2.11.1
pyflakes==3.1.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
virtualenv==20.24.5
